proposed text on network-based authentication
The subscriber is not necessarily  the user. Network-based authentication identifies the subscription. A child of the subscriber using their phone which has a contract owned by a parent might not be legally allowed to decide about consent.
Network-based authentication is only a second factor to user authentication and user consent.